refactor: MESSAGE 위치 변경
