+++ src/ts/ui/MainUI/CoinManagementUI/CoinChargeUI.ts
@@ -1,4 +1,4 @@
-import { MESSAGE } from '../../../constants';
+import { MESSAGE } from '../../../constants/message';
 import { showSnackbar } from '../../../utils';
 import { $ } from '../../../utils/dom';
 import { viewPainter } from '../../ViewPainter';
