+++ src/ts/ui/MainUI/ProductManagementUI/ProductInventoryUI.ts
@@ -1,4 +1,4 @@
-import { MESSAGE } from '../../../constants';
+import { MESSAGE } from '../../../constants/message';
 import { showSnackbar } from '../../../utils';
 import { $, $$, replaceHTML } from '../../../utils/dom';
 import { viewPainter } from '../../ViewPainter';
