+++ src/ts/ui/MainUI/ProductPurchaseUI/PurchaseCashChargeUI.ts
@@ -1,4 +1,4 @@
-import { MESSAGE } from '../../../constants';
+import { MESSAGE } from '../../../constants/message';
 import { showSnackbar } from '../../../utils';
 import { $ } from '../../../utils/dom';
